rename rest_api.py -> htm_rest_api.py
